Refactor dataQuestion
